fix(mycontexts): removeRole now handles multiple instances
The RecentContexts Component removes excess roles from
RecentContexts (the System role). Previously, it iterated over the
instances to be removed. This caused a perspective refresh, leading to
a new iteration over the rest of the roles to be removed. Consequently,
many role instances were removed more than once and that errors.