fix(core): fixed DomeinFile loading while compiling an arc file
When compiling in phase three, we deal with
Readable identifiers. So when we look up an identifier from the model
under compilation, we don't want to switch to Stable identifiers
and we don't want to load its defining model from the Repository.
Instead, we want to rely on the version in cache as we have put the
Readable record in it on starting phase three.