fix(core): improved type switching and arc compilation now is complete
We now keep sidecars in PhaseTwoState. This improves switching from
readable to stable identifiers and vv. In PhaseThree, during compilation
of an arc file, we need to switch to readable in many places. Added
many instances of StableToReadable. The readable2stable function is
now available to be used outside module Perspectives.Representation.Class.PersistentType, too.